fix(Tests): 🐛 Prepare tests file
