+++ package.json
@@ -17,9 +17,14 @@
     "alien",
     "coriolis",
     "tales from the loop",
+    "things from the flood",
     "vaesen",
     "twilight 2000",
+    "twilight 2000 4e",
     "t2k",
+    "t2k4e",
+    "tw2k",
+    "tw2k4e",
     "blade runner",
     "free league publishing",
     "free league",
@@ -27,6 +32,7 @@
     "dice",
     "roll",
     "discord",
+    "discordjs",
     "bot"
   ],
   "readme": "https://github.com/Stefouch/sebedius-yearzero-discord-bot/blob/master/README.md",
@@ -45,15 +51,15 @@
   "scripts": {
     "start": "node .",
     "lint": "eslint .",
-    "lint:fix": "echo not yet",
+    "lint:fix": "eslint --ext .js --fix .",
     "upd": "npm outdated & npx npm-check -u",
     "husky:install": "npx husky install",
     "test": "mocha ./test/*.test.js --exit",
     "test2": "mocha ./test/roll.test.js ./test/lang.test.js ./test/client.test.js --exit",
-    "test-roll": "mocha ./test/roll.test.js --exit",
-    "test-lang": "mocha ./test/lang.test.js --exit",
-    "test-client": "mocha ./test/client.test.js --exit",
-    "test-db": "mocha ./test/db.test.js --exit",
+    "test:roll": "mocha ./test/roll.test.js --exit",
+    "test:lang": "mocha ./test/lang.test.js --exit",
+    "test:client": "mocha ./test/client.test.js --exit",
+    "test:db": "mocha ./test/db.test.js --exit",
     "precommit-msg": "devmoji -e --lint",
     "pm2:watch": "pm2 start bot.js --watch & pm2 log",
     "register": "node register.js"
